Fixed bugs and added a pop-up window for solutions.
DETAILS==

Fixed a bug with the board in which words would no longer be added to the
users found words list after the same word was input twice.
Added a pop-up window within the OnePlayer Activity that displays all the
possible words within the given board.

AFFECTED AREAS==

Most changes revolved around code associated with the OnePlayer activity
and several new files were created in order to generate the pop-up window.

TESTING==

Manual testing was performed in order to verify the bug fix.